[FIX] Trade Hoarding Limits
